Provide a secure way to set parameters
This is only in the direction of securely setting parameters.

Given the parameters are set on the command line directly, the security
sensitive values should probably be moved to an actually k8s secret so
they can be passed as secret environment values.

That said, this does demonstrate how not to hard-code the username and
password for our database user.